requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.2
mysqlclient==1.3.10
Flask-Migrate==2.0.4
google-cloud-language==1.0.2
